Fix deploy script
It looks like Deno's stdlib was restructured here: https://github.com/denoland/deno_std/pull/813.

Some files/modules were marked deprecated and re-exported parts of the new modules. In this case, `std/io/bufio.ts` was deprecated in favor of `std/io/buffer.ts`. As of `std@0.159.0`, `std/io/bufio.ts` was removed completely.

To prevent similar issues in the future, it would be best to pin to a specific version using one of the following:
- Versioned import i.e. `import * as bufio from "https://deno.land/std@1.58.0/io/bufio.ts"
- Importmap
- Lock file
- Vendoring i.e. `deno vendor`