Enable `skipLibCheck` for React Native tsconfig
`skipLibCheck` makes it so that `tsc` will not typecheck library declarations. It was set in the previous TypeScript template, overriding the base @tsconfig/react-native, to avoid cases of duplicate identifiers between global types in React Native vs Node. I didn't enable it initially in the RN copy of the template since it reduces safety, didn't seem to cause issues in an example app, and wasn't the right place for the change compared to the base config.

In a more full repo I ran into the same issue. I was a bit confused, since specifying explicit `types` in `compilerOptions` would mean we are not brining it in for our own code. Telling tsc to ignore the Node types also isn't viable. Enabling `explainFiles`, I was able to see why it was being pulled in despite not being told to. In my specific case, a `devDependency` on a globbing library for a script had `/// <reference types="node" />` in a declaration generated by the TypeScript compiler itself, which not only brought in the Node types being exported, but also created conflicts in the global namespace. Quite a few libraries I tried seem to do this, so for now it seems like it would make sense to replicate the behavior of the previous template, and be permissive of libraries which have a dependency on Node typings.